+++ client-report/src/components/correlationMatrix/matrix.js
@@ -3,6 +3,8 @@
 import React from "react";
 // import { connect } from "react-redux";
 import Flex from "../framework/flex";
+import * as d3 from "d3";
+import { interpolatePuOr } from 'd3-scale-chromatic';
 
 import Radium from "radium";
 import _ from "lodash";
@@ -38,7 +40,7 @@ class Matrix extends React.Component {
     return (
       <g>
         <rect
-          fill={d3.interpolatePuOr(scale(comment))}
+          fill={interpolatePuOr(scale(comment))}
           onMouseEnter={() => {
             return this.onMouseEnterCell(row, column, comment);
           }}
@@ -176,7 +178,7 @@ class Matrix extends React.Component {
               x={300}
               y={40}
               textAnchor={"middle"}
-              fill={d3.interpolatePuOr(scale(this.state.mouseOverCorrelation))}
+              fill={interpolatePuOr(scale(this.state.mouseOverCorrelation))}
               style={{
                 fontFamily: globals.sans,
                 fontSize: 18